requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
yfinance>=0.2.0
pyarrow>=14.0.0
pandas-ta>=0.3.14b
pytest>=7.4.0
plotly>=5.18.0
kaleido>=0.2.1
supabase>=2.0.0
python-dotenv>=1.0.0